Gladius Posterioris
And yea, the great sord of fire rent the skies with great rending and sang us the thirteen songs of the posterior.

Lo, let it be seen that by the hand of the sord and the rear of the sord do we have more sound that descends from upon high the mountaintops and praise it with great praise.

Adds 13 new farts.