Ported Clothing And Resurrected Cosmetics (#988)
* Ports some neat clothing from Citadel,
Skyrat, TG. Also saves some cosmetics
from the hellscape that is the unused
Supporter/Donator meta store section
and re-adds them to the regular meta store. Some of the new/resurrected
stuff is avalible in vendors, but not all of it. All of it although can
purchased in the metacoin store

* I forgot the phantom glasses and driscoll boots
also added bloodred pajamas to clothes mate contraband

* fix neck items, add boots

---------

Co-authored-by: DESKTOP-M4GD565\LME <polygobttv@gmail.com>